safari-600.7-branch/LayoutTests/platform/ios-sim/fast/events/wheelevent-basic-expected.txt

TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP TOP
Scroll mouse wheel over here
Scroll mouse wheel over here
Scroll mouse wheel over here
Scroll mouse wheel over here
Scroll mouse wheel over here
Scroll mouse wheel over here
END END END END END END END END END END END END END END
Tests the basic functionality of the standard wheel event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WheelEvent.prototype.__proto__ is MouseEvent.prototype
PASS WheelEvent.DOM_DELTA_PIXEL is 0x00
PASS WheelEvent.DOM_DELTA_LINE is 0x01
PASS WheelEvent.DOM_DELTA_PAGE is 0x02
PASS window.onwheel is null
PASS document.onwheel is null
PASS testDiv.onwheel is null
FAIL deltaX > 0 should be true. Was false.
PASS deltaX is expectedDeltaX
FAIL deltaY > 0 should be true. Was false.
PASS deltaY is expectedDeltaY
PASS successfullyParsed is true

TEST COMPLETE
